fixes up get_agent_accounts
