+++ elfpy/bots/bot_info.py
@@ -37,7 +37,7 @@ class BotInfo:  # FIXME: Rename to `BotConfig` when we remove evm_bots
     index: int | None = None  # TODO: Make this required when we remove evm_bots
     name: str = "BoringBotty"
     budget: Budget = Budget()
-    number_of_bots: int = 1
+    number_of_agents: int = 1
     init_kwargs: dict = field(default_factory=dict)
     # TODO: Remove below once we remove evm_bots
     trade_chance: float = 0.8
